added refactor delete task

+++ src/services/task/delete.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from "@presentation";
 import { ClientTypesAdapter, IClientRepository, IDeleteTaskService, TaskRequestDto } from "@protocols";
 
 export class DeleteTaskService implements IDeleteTaskService {
@@ -6,7 +7,10 @@ export class DeleteTaskService implements IDeleteTaskService {
   constructor(private clientRepository: IClientRepository) { }
 
   public async delete(request: TaskRequestDto): Promise<void> {
-    await this.clientRepository.get(this.params(request));
+    const data = await this.clientRepository.get(this.params(request));
+    if (!data.Item) {
+      throw new NotFoundError('Task not found with given identifiers');
+    }
     await this.clientRepository.delete(this.params(request));
   }
